requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
